added tryingtoenter boolean but there are still some issues
sometimes two nodes are in the critical section at once....
Could be something to do with the way the next element is selected from the queue if the timestamp is the same? But I am not sure that is even possible. And if it is, the next node is selected from only one node anyhow - the node leaving. So what is going on? Not sure.